Movendo get_handler_original_typehints para asyncworker.typing
